Rename: assets 폴더로 이동
