+++ src/view/InputView.js
@@ -1,6 +1,6 @@
 const MissionUtils = require("@woowacourse/mission-utils");
 const bridgeGameController = require('../controller/BridgeGameController');
-const MESSAGE = require("../Messages");
+const MESSAGE = require("../assets/Messages");
 /**
  * 사용자로부터 입력을 받는 역할을 한다.
  */
