+++ src/view/OutputView.js
@@ -1,5 +1,5 @@
 const MissionUtils = require('@woowacourse/mission-utils')
-const MESSAGE = require('../Messages');
+const MESSAGE = require('../assets/Messages');
 /**
  * 사용자에게 게임 진행 상황과 결과를 출력하는 역할을 한다.
  */
